Update scoring system to calc losing score

+++ app/src/main/java/io/github/patpatchpatrick/nflseasonsim/season_resources/Team.java
@@ -20,14 +20,14 @@ public class Team {
 
     private String mName;
     private double mElo;
-    private int mOffRating;
-    private int mDefRating;
+    private double mOffRating;
+    private double mDefRating;
     private int mCurrentWins;
     private int mCurrentLosses;
     private int mCurrentDraws;
     private int mDivision;
 
-    public Team(String name, double elo, int offRating, int defRating, int division){
+    public Team(String name, double elo, double offRating, double defRating, int division){
 
         //Inject team with dagger to get contentResolver
         DaggerApplication.getAppComponent().inject(this);
@@ -66,7 +66,7 @@ public class Team {
     }
 
     public double getELO() {
-        return (double) mElo;
+        return mElo;
     }
 
     public void  setELO( double elo ) {
